+++ projects/common/lib/components/address-validator/address-validator.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed, getTestBed, fakeAsync, tick } from '@
 import { FormGroup, FormBuilder } from '@angular/forms';
 import { AddressValidatorComponent } from './address-validator.component';
 import { FormsModule } from '@angular/forms';
-import { TypeaheadModule, TypeaheadMatch } from 'ngx-bootstrap';
+import { TypeaheadModule, TypeaheadMatch } from 'ngx-bootstrap/typeahead';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { of, throwError } from 'rxjs';
@@ -54,14 +54,16 @@ describe('AddressValidatorComponent', () => {
       City: 'Victoria',
       DeliveryAddressLines: '784 Yates St',
       Country: 'Canada',
-      Province: 'British Columbia'
+      Province: 'British Columbia',
+      PostalCode: 'V1V1V1'
     },
     {
       AddressComplete: '784 Young Rd, Kelowna, BC',
       City: 'Kelowna',
       DeliveryAddressLines: '784 Young Rd',
       Country: 'Canada',
-      Province: 'British Columbia'
+      Province: 'British Columbia',
+      PostalCode: 'V2V2V2'
     }
   ];
 
@@ -168,6 +170,93 @@ describe('AddressValidatorComponent', () => {
 
     component.onSelect(mockEvent);
     expect(component.select.emit).toHaveBeenCalledWith(jasmine.any(Address));
+  }));
+
+  it('should set `hasError` to true when encountering an error', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockError = {};
+    component.onError(mockError);
+    expect(component.hasError).toBe(true);
+  }));
+
+  it('should set `isTypeaheadLoading` to true when passed in `onLoading`.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    component.hasError = true;
+    component.onLoading(true);
+    expect(component.isTypeaheadLoading).toBe(true);
+    expect(component.hasError).toBe(false);
+  }));
+
+  it('should set `hasNoResults` to true when passed in `onNoResults`.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    component.hasError = true;
+    component.onNoResults(true);
+    expect(component.hasNoResults).toBe(true);
+    expect(component.hasError).toBe(true);
+    component.onNoResults(false);
+    expect(component.hasNoResults).toBe(false);
+    expect(component.hasError).toBe(false);
+  }));
+
+  it('should emit `address` with values.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockSelectedItem = {
+      item: yatesResponse[0]
+    } as TypeaheadMatch;
+    spyOn(component.select, 'emit').and.returnValue(null);
+    component.selectedAddress = false;
+    component.onSelect(mockSelectedItem);
+    expect(component.select.emit).toHaveBeenCalledWith(jasmine.any(Object));
+    expect(component.selectedAddress).toBe(true);
+  }));
+
+  it('should set `search to address selected when `populateAddressOnSelect` is true.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockSelectedItem = {
+      item: yatesResponse[0]
+    } as TypeaheadMatch;
+    spyOn(component.select, 'emit').and.returnValue(null);
+    component.populateAddressOnSelect = true;
+    component.onSelect(mockSelectedItem);
+    expect(component.search).toBe(yatesResponse[0].DeliveryAddressLines);
+  }));
+
+  it('should handle keyUp.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockKeyboardEvent = {
+      keyCode: 1
+    } as KeyboardEvent;
+    component.selectedAddress = true;
+    component.onKeyUp(mockKeyboardEvent);
+    expect(component.selectedAddress).toBe(false);
+  }));
+
+  it('should handle onBlur.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockEvent = {};
+    const mockSearch = 'Test';
+    spyOn(component, '_onTouched').and.returnValue(null);
+    spyOn(component, '_onChange').and.returnValue(null);
+    component.search = mockSearch;
+    component.onBlur(mockEvent);
+    expect(component._onTouched).toHaveBeenCalled();
+    expect(component._onChange).toHaveBeenCalledWith(component.search);
+  }));
+
+  it('should write value.', fakeAsync(() => {
+    fixture = TestBed.createComponent(AddressValidatorComponent);
+    component = fixture.componentInstance;
+    const mockSearch = 'Test';
+    component.search = null;
+    component.writeValue(mockSearch);
+    expect(component.search).toBe(mockSearch);
   }));
 });