+++ projects/common/lib/components/file-uploader/file-uploader.component.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { FormsModule, ControlContainer, NgForm } from '@angular/forms';
 import { FileUploaderComponent } from './file-uploader.component';
 import {ThumbnailComponent} from '../thumbnail/thumbnail.component';
-import {ModalModule} from 'ngx-bootstrap';
+import {ModalModule} from 'ngx-bootstrap/modal';
 import { RouterTestingModule } from '@angular/router/testing';
 import * as moment from 'moment';
 import {HttpClientModule} from '@angular/common/http';